requirements.txt
pytest
sphinx
sphinx_rtd_theme
black
loguru
hydra-core
streamlit
pandas
matplotlib
numpy